pyproject.toml
[tool.poetry]
name = "audio-testing"
version = "0.1.0"
description = ""
authors = ["noahdahlman <noah.dahlman@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pydub = "^0.25.1"
ipfshttpclient = "^0.7.0"
autopep8 = "^1.6.0"
python-dotenv = "^0.19.2"
wget = "^3.2"
fastapi = {extras = ["all"], version = "^0.74.0"}

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
